Add basic camera loop
Adds simple loop for capturing video stream using OpenCV.
Video frames can be sent to a computer vision model for
distraction detection.

The camera loop displays the user's camera to them in a
pop-up window. This is so we can visually see when the app
is detecting the user is distracted and also get feedback
on how fast the CV model runs. This will also be good when
demo-ing the functionality.