Add detection threshold to camera loop
Implements a circular buffer to find distraction rate
over the last few seconds. If the distraction rate is
above a specific threshold, we can start recording.

This is needed to make the application more robust to
outliers from the model predictions. It is also useful
for preventing very momentary distractions from triggering
a summary pop-up.